[nj] refactor index page of Blog
Remember that Object (which is returned by `getPage` or `site.__matter__`) doesn't guarantee to be properly sorted.

If you want to have specific order, you need to sort entries based on something. For example, date. Or you can pass to page prop `order` with needed value and sort pages based on it.